+++ hyperdbg/hprdbghv/pch.h
@@ -108,6 +108,7 @@
 #include "hooks/ModeBasedExecHook.h"
 #include "interface/Callback.h"
 #include "features/DirtyLogging.h"
+#include "features/Tracing.h"
 #include "features/CompatibilityChecks.h"
 
 //